Uploaded GetHost.js
This is used to get a urls name like www.google.com but without the www. or .com things,
I dont see any use of it for anyone else other than me for my IsUrl.Website function check out IsUrl.js
As i just modified the GetHost function for my needs in seeing if a url is a website.